style: 메인페이지와 상세페이지 css다룸
