+++ mainModules/util.js
@@ -1,0 +1,5 @@
+export function strToHtml(str) {
+    const parser = new DOMParser();
+    const result = parser.parseFromString(str, "text/html");
+    return result;
+}
