+++ spec.css
@@ -1,7 +1,11 @@
 * {
-  margin: 0;
+  margin: 2px;
   padding: 0;
   box-sizing: border-box;
+}
+
+.full-page {
+  padding-left: 20px;
 }
 
 html {
@@ -10,16 +14,20 @@ html {
 
 body {
   font-family: Arial, sans-serif;
-  background-color: #000;
-  color: #fff;
-  padding: 20px;
+  background-color: #f6f3ef;
+  /* 밝은 배경색 */
+  /* padding-top: 20px; */
+  background-image: url("https://i.ibb.co/MRqx3pc/alex-koch-v-WBGy7-Bc-AMk-unsplash.jpg");
+  background-size: cover;
+  background-color: repeat-y;
+  /* 이미지를 가운데로 정렬 */
 }
 
 header {
   top: 0;
   left: 0;
   width: 100%;
-  background-color: rgba(39, 39, 42, 0.9);
+  background-color: #e0d1ff;
   border-bottom: none;
   position: fixed;
   z-index: 99;
@@ -37,55 +45,65 @@ header {
   margin: 0 auto 0 0;
 }
 
-.pageTitle > a {
+.pageTitle>a {
   font-size: 4rem;
-  color: #fff;
+  color: #c993f3;
   font-weight: 600;
-
+  text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5);
+  -webkit-text-stroke: 0.5p#2c240200;
   text-decoration: none;
 }
 
-.pageTitle > a > span {
-  color: rgb(248, 236, 11);
+.pageTitle>a>span {
+  color: #f2e1ff;
 }
 
-.headerForm > a {
+.headerForm>a {
   padding-right: 10px;
 }
 
-.headerForm > p {
-  color: #fff;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
+.headerForm>p {
+  color: #c993f3;
+  font-weight: bold;
   font-size: 1.2rem;
   letter-spacing: 1.5px;
+  padding-right: 15px;
 }
 
 .container {
+  margin-top: 120px;
   display: flex;
   flex-direction: row;
 }
+
 .poster {
   margin-right: 20px;
 }
+
 .poster img {
   width: 200px;
   height: 300px;
 }
+
 .details {
   flex: 1;
 }
+
 h1 {
   font-size: 28px;
   margin-bottom: 10px;
 }
+
 p {
   font-size: 16px;
   line-height: 1.5;
   margin-bottom: 10px;
 }
+
 .genres {
   margin-bottom: 20px;
 }
+
 .genres span {
   background-color: #555;
   padding: 5px 10px;
@@ -93,28 +111,37 @@ p {
   font-size: 14px;
   border-radius: 4px;
 }
+
 .trailer {
   width: 100%;
   margin-bottom: 20px;
   float: right;
 }
+
 .trailer #movieDetails {
   float: right;
   width: 40%;
   height: 500px;
+  padding-left: 20px;
 }
 
 .trailer #titleSection {
-    font-size: 5em;
-    padding-bottom: 50px;
+  font-size: 5em;
+  padding-bottom: 50px;
 }
-.blink{
-    animation: blink 1s step-end infinite;
+
+#movieOverview {
+  font-size: 25px;
 }
+
+.blink {
+  animation: blink 1s step-end infinite;
+}
+
 @keyframes blink {
-50% {
+  50% {
     opacity: 0;
-   }
+  }
 }
 
 .trailer #movieTrailer {
@@ -122,59 +149,236 @@ p {
   width: 60%;
   height: 500px;
 }
+
 .trailer iframe {
   width: 100%;
   height: 100%;
 }
 
+.container h2 {
+  margin-top: 20%;
+  font-weight: bold;
+  font-size: 18px;
+}
+
+.cast-section {
+  margin-top: 30px;
+}
+
+.actor {
+  max-height: 320px;
+  display: inline-block;
+}
+
+
+.actor h3 {
+  padding-top: 5%;
+  font-weight: bold;
+  font-size: 18px;
+}
+
+.actor p {
+  padding-top: 3%;
+}
+
+
 .cast-card {
   /* div class명을 cast-card로 변경함 */
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  align-items: center;
-  margin-bottom: 20px;
+  height: auto;
+  margin-top: 1%;
+  margin-bottom: 50px;
+  width: 1450px;
+  overflow-y: scroll;
+  overflow: scroll;
+  white-space: nowrap;
 }
+
+.cast-card::-webkit-scrollbar {
+  display: none;
+  /* Chrome/Safari/Opera에서 스크롤바 숨기기 */
+}
+
+
 .cast-card img {
   /* cast-card 안의 img를 지정함 */
   width: auto;
-  height: 200px;
+  height: 250px;
   /* width, heigt 값만 변경함 */
   /* 가로 세로 길이를 지정해버리면 img가 강제로 그 사이즈에 맞춰지는 듯 -> 찌그러져 보임 .. */
   background-color: #555;
-  margin-right: 10px;
+  margin-right: 20px;
+}
+
+.actor {
+  max-width: 160px;
+  word-wrap: break-word;
+  /* word-break:break-all; */
+  margin-right: 20px;
+}
+
+.comment-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-left: 450px;
 }
 
 .comments {
-  margin-bottom: 20px;
+  margin-top: 30px;
+  margin-bottom: 80px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
+
 .comments .comment {
   margin-bottom: 10px;
   padding: 10px;
-  background-color: #222;
+  background-color: #e0d1ff;
+  border: solid 2px #c993f3;
+  width: 500px;
+  border-radius: 10px;
+  margin-left: 15px;
 }
+
 .comments .comment p {
-  margin: 0;
+  margin: 2px;
 }
+
 .comments .comment .username {
   color: #aaa;
   font-weight: bold;
 }
+
 .comments .comment .timestamp {
   color: #aaa;
   font-size: 12px;
+  padding-bottom: 20px;
 }
+
+.buttons {
+  padding-top: 20px;
+}
+
 .comment-form {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin-top: 10%;
   margin-bottom: 20px;
 }
-.comment-form textarea {
-  width: 100%;
-  height: 100px;
+
+.input-container {
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
 }
+
+.rating-input {
+  margin-right: 10px;
+  width: 130px;
+  height: 30px;
+  padding-left: 10px;
+  border-radius: 6px;
+  border: solid 2px #1E90FF;
+}
+
+.input-container input {
+  height: 30px;
+  width: 210px;
+  margin-right: 10px;
+  padding: 10px;
+  box-sizing: border-box;
+  border: solid 2px #1E90FF;
+  border-radius: 5px;
+  font-size: 16px;
+  resize: none; 
+
+}
+
+.textarea-container {
+  margin-bottom: 10px;
+  width: 1000px;
+  
+}
+
+.comment-form textarea {
+  width: 57%;
+  height: 100px;
+  padding: 10px;
+  box-sizing: border-box;
+  border: solid 2px #1E90FF;
+  border-radius: 5px;
+  font-size: 16px;
+  resize: none; 
+}
+
 .comment-form button {
-  background-color: #555;
-  color: #fff;
-  padding: 5px 10px;
+  background-color: #e0d1ff;
+  border: solid 2px #c993f3;
+  color: #000000;
+  padding: 5px 70px;
   border: none;
   cursor: pointer;
+  border-radius: 20px;
+  height: 30px;
+}
+
+.comment-center button {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-left: 200px;
+}
+
+footer {
+  position: absolute;
+  width: 100%;
+  background-color: #e0d1ff;
+  text-align: center;
+}
+
+.team-members {
+  margin: 20px;
+  margin-bottom: 20px;
+}
+
+.team-members span {
+  font-weight: bold;
+  font-size: 20px;
+}
+
+.team-members ul {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
+
+.team-members li {
+  margin-top: 13px;
+}
+
+.github-link {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding-top: 10px;
+}
+
+.github-link a {
+  text-decoration: none;
+  display: inline-flex;
+  align-items: center;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.github-link #logo{
+  transition-duration: 0.2s;
+}
+
+.github-link #logo:hover {
+  transform: scale(1.3);
+  transition-duration: 0.5s;
 }
